tumbo/requirements.txt
APScheduler==3.0.4
bumpversion==0.5.3
bunch==1.0.1
configobj==5.0.6
Django==1.7.10
django-compressor==1.5
django-debug-toolbar==1.2.2
django-extensions==1.4.6
django-gravatar2==1.1.4
django-redis==3.8.3
django-redis-metrics==1.0.3
djangorestframework==2.4.3
django-rest-swagger==0.3.4
django-solo==1.1.0
docker-compose>=1.5.2
docopt==0.6.2
dropbox==1.6
ecdsa==0.10
docker-py>=1.8.1
-e git+https://github.com/rpalacios/django-sequence-field.git#egg=django_sequence_field-master
gevent==1.0.1
greenlet==0.4.5
gunicorn==19.1.1
jsonfield==0.9.22
Markdown==2.6.6
mock==1.0.1
ndg-httpsclient==0.3.2
newrelic==2.58.2.45
oauthlib==0.7.1
objgraph==2.0.1
opbeat==3.1.1
pika==0.9.14
psutil==3.3.0
psycopg2==2.6.1
pyasn1==0.1.8
pyflakes==0.8.1
Pygments==2.1.3
pyOpenSSL==0.14
pytz==2014.9
redis==2.10.3
requests==2.7.0
sh==1.09
shortuuid==0.4
SQLAlchemy==1.0.12
tabulate==0.7.5
